requirements.txt
absl-py==1.4.0
accelerate==0.33.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.3
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bitsandbytes==0.42.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distro==1.9.0
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring-parser==0.15
etils==1.7.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.1
idna==3.6
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.11
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.30.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
peft==0.7.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.1.0
plotly==5.23.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.40.6
setproctitle==1.3.3
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-datasets==4.9.4
tensorflow-estimator==2.15.0
tensorflow-gcs-config==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.23.0
termcolor==2.4.0
tf-keras==2.15.0
threadpoolctl==3.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.43.1
triton==2.2.0
trl==0.9.6
typing_extensions==4.9.0
tyro==0.6.6
tzdata==2023.4
urllib3==2.1.0
wandb==0.17.5
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.11
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
